fix(ci): keep schema generation native-free
